patch 7.4.1587 Problem:    Compiler warnings with 64 bit compiler. Solution:   Add type casts. (Mike Williams)

+++ src/ex_cmds2.c
@@ -3272,7 +3272,7 @@ do_in_runtimepath(
 
     if ((done == FAIL || (flags & DIP_ALL)) && (flags & DIP_START))
     {
-	len = STRLEN(start_dir) + STRLEN(name);
+	len = (int)(STRLEN(start_dir) + STRLEN(name));
 	s = alloc(len);
 	if (s == NULL)
 	    return FAIL;
@@ -3283,7 +3283,7 @@ do_in_runtimepath(
 
     if ((done == FAIL || (flags & DIP_ALL)) && (flags & DIP_OPT))
     {
-	len = STRLEN(opt_dir) + STRLEN(name);
+	len = (int)(STRLEN(opt_dir) + STRLEN(name));
 	s = alloc(len);
 	if (s == NULL)
 	    return FAIL;
